+++ src/logging/logger.ts
@@ -1,12 +1,12 @@
-import winston, { log, Logger } from "winston";
+import winston from "winston";
 import { env } from "../config";
 
-const { combine, timestamp, json, errors, printf } = winston.format;
+const { combine, timestamp, errors, prettyPrint } = winston.format;
 
 const logger = winston.createLogger({
-	level: env.LOG_LEVEL,
-	format: combine(errors({ stack: true }), timestamp(), json()),
-	transports: [new winston.transports.Console()],
+  level: env.LOG_LEVEL,
+  format: combine(errors({ stack: true }), timestamp(), prettyPrint()),
+  transports: [new winston.transports.Console()],
 });
 
 export default logger;
