+++ src/services/security/lib/encryption/symmetric/standard/handler.ts
