+++ src/services/security/lib/key-management/key-serializer.ts
