+++ src/services/security/symmetric-encryption.service.ts
